requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
confection==0.1.5
cymem==2.0.8
cytoolz==0.12.3
dataclasses-json==0.6.7
Deprecated==1.2.14
diff-match-patch==20230430
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.4
django-activity-stream==2.0.0
django-cors-headers==4.4.0
django-import-export==4.1.0
django-user-activity-log==0.0.15
django-user-visit==2.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
environs==9.5.0
faiss-cpu==1.8.0
filelock==3.15.4
Flask==3.0.3
floret==0.10.5
frozenlist==1.4.1
fsspec==2024.6.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
import-export==0.3.1
itsdangerous==2.2.0
jellyfish==1.0.4
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.9
langchain-community==0.2.10
langchain-core==0.2.23
langchain-experimental==0.0.62
langchain-google-genai==1.0.6
langchain-postgres==0.0.9
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langcodes==3.4.0
langsmith==0.1.81
language_data==1.2.0
llama-cloud==0.0.9
llama-index==0.10.55
llama-index-agent-openai==0.2.8
llama-index-cli==0.1.12
llama-index-core==0.10.55
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.5
llama-index-legacy==0.9.48
llama-index-llms-ollama==0.1.5
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.30
llama-index-readers-llama-parse==0.1.6
llama-parse==0.4.4
llamaindex-py-client==0.1.19
llvmlite==0.43.0
mariadb==1.1.10
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
milvus-lite==2.4.8
minijinja==2.0.1
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
ollama==0.3.0
openai==1.35.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pgvecto-rs==0.1.5
pgvector==0.2.5
pillow==10.3.0
preshed==3.0.9
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymilvus==2.4.4
pyparsing==3.1.2
pypdf==4.2.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.0.1
setuptools==70.1.0
setuptools-rust==1.9.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
srsly==2.4.8
striprtf==0.0.26
sympy==1.12.1
tablib==3.5.0
tenacity==8.4.1
textacy==0.13.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tqdm==4.66.4
transformers @ git+https://github.com/huggingface/transformers@74a207404e8d4524d1fdc4aa23789694f9eef347
triton==2.3.1
typer==0.12.3
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.2
user-agents==2.2.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
